shell: stop using GNU coreutils by default
I've been putting the GNU coreutils at the head of my path for a long
time and haven't had any real issues until moving to an M1 mac. The
brew installed version of the utilities include `uname` but it's not a
universal binary. When it's at the head of the path it will always
return "arm64" even when running in an x86_64 environment.