nvim: injection for probable json in raw string
go.nvim has injections for literals that are assigned to consts and vars
of string literals with names that include [Jj]son but, in my
experience, that's pretty uncommon.

This adds an injection that highlights raw string literals that start
with `{` or '[` as JSON. It's not perfect but will probably do what I
want 99% of the time.