nvim: try to configure debug adapter for java
This attempt does have an issue where stepping into source associated
with a dependency causes an error with the nvim-dap-virtual-text where
the filetype of the buffer cannot be detected.

The filetype doesn't seem to be set when nvim-dap-virtual-text runs, so
it tries to discover it from the file extension but the .class extension
is associated with `stata` instead of `java`. Not sure how to fix this
yet.